feat(organizations): improve UI & UX (#387)

+++ lib/atomic/organizations.ex
@@ -35,6 +35,25 @@ defmodule Atomic.Organizations do
   end
 
   @doc """
+  Returns the list of organizations members.
+
+  ## Examples
+
+      iex> list_organizations()
+      [%Membership{}, ...]
+
+  """
+  def list_organizations_members(%Organization{} = organization) do
+    from(m in Membership,
+      where: m.organization_id == ^organization.id and m.role != :follower,
+      join: u in User,
+      on: u.id == m.user_id,
+      select: u
+    )
+    |> Repo.all()
+  end
+
+  @doc """
   Returns the list of organizations followed by an user.
 
   ## Examples
@@ -365,7 +384,12 @@ defmodule Atomic.Organizations do
 
   """
   def delete_membership(%Membership{} = membership) do
-    Repo.delete(membership)
+    organization =
+      from(o in Organization, where: o.id == ^membership.organization_id) |> Repo.one()
+
+    if !(organization.name == "CeSIUM" and membership.role == :follower) do
+      Repo.delete(membership)
+    end
   end
 
   @doc """
